Move up the edit method.
