+++ newIDE/app/src/ObjectEditor/Editors/SpriteEditor/index.js
@@ -4,7 +4,10 @@ import { t } from '@lingui/macro';
 import { I18n } from '@lingui/react';
 import { type I18n as I18nType } from '@lingui/core';
 import * as React from 'react';
-import SpritesList, { addAnimationFrame } from './SpritesList';
+import SpritesList, {
+  addAnimationFrame,
+  applyPointsAndMasksToSpriteIfNecessary,
+} from './SpritesList';
 import IconButton from '../../../UI/IconButton';
 import FlatButton from '../../../UI/FlatButton';
 import RaisedButton from '../../../UI/RaisedButton';
@@ -35,14 +38,27 @@ import GDevelopThemeContext from '../../../UI/Theme/GDevelopThemeContext';
 import useAlertDialog from '../../../UI/Alert/useAlertDialog';
 import { getMatchingCollisionMask } from './CollisionMasksEditor/CollisionMaskHelper';
 import {
+  copySpritePoints,
+  copySpritePolygons,
   getCurrentElements,
   getTotalSpritesCount,
 } from './Utils/SpriteObjectHelper';
 import Edit from '../../../UI/CustomSvgIcons/Edit';
 import { groupResourcesByAnimations } from './AnimationImportHelper';
 import { applyResourceDefaults } from '../../../ResourcesList/ResourceUtils';
+import { ExternalEditorOpenedDialog } from '../../../UI/ExternalEditorOpenedDialog';
+import {
+  type ResourceExternalEditor,
+  type EditWithExternalEditorReturn,
+} from '../../../ResourcesList/ResourceExternalEditor';
+import { showErrorBox } from '../../../UI/Messages/MessageBox';
 
 const gd: libGDevelop = global.gd;
+
+const removeExtensionFromFileName = (fileName: string) => {
+  const dotIndex = fileName.lastIndexOf('.');
+  return dotIndex < 0 ? fileName : fileName.substring(0, dotIndex);
+};
 
 const DragSourceAndDropTarget = makeDragSourceAndDropTarget(
   'sprite-animations-list'
@@ -108,6 +124,7 @@ export default function SpriteEditor({
   onObjectUpdated,
   isAnimationListLocked = false,
 }: SpriteEditorProps) {
+  const [externalEditorOpened, setExternalEditorOpened] = React.useState(false);
   const [pointsEditorOpen, setPointsEditorOpen] = React.useState(false);
   const [advancedOptionsOpen, setAdvancedOptionsOpen] = React.useState(false);
   const [
@@ -513,7 +530,151 @@ export default function SpriteEditor({
 
       if (onObjectUpdated) onObjectUpdated();
     },
-    [resourceManagementProps, resourceSources, addAnimations, forceUpdate, onObjectUpdated, project]
+    [
+      resourceManagementProps,
+      resourceSources,
+      addAnimations,
+      forceUpdate,
+      onObjectUpdated,
+      project,
+    ]
+  );
+
+  const adaptCollisionMaskIfNeeded = React.useCallback(() => {
+    // If the first sprite of the first animation is updated,
+    // we update the automatic collision mask of the object,
+    // if the option is enabled.
+    if (spriteConfiguration.adaptCollisionMaskAutomatically()) {
+      onCreateMatchingSpriteCollisionMask();
+    }
+  }, []);
+
+  const editWith = React.useCallback(
+    async (
+      i18n: I18nType,
+      externalEditor: ResourceExternalEditor,
+      direction: gdDirection,
+      animationIndex: number,
+      directionIndex: number
+    ) => {
+      const resourceNames = mapFor(0, direction.getSpritesCount(), i => {
+        return direction.getSprite(i).getImageName();
+      });
+      const animation = spriteConfiguration.getAnimation(animationIndex);
+      const animationName = animation.getName();
+
+      try {
+        setExternalEditorOpened(true);
+        const editResult: EditWithExternalEditorReturn | null = await externalEditor.edit(
+          {
+            project,
+            i18n,
+            getStorageProvider: resourceManagementProps.getStorageProvider,
+            resourceManagementProps,
+            resourceNames,
+            extraOptions: {
+              singleFrame: false,
+              fps:
+                direction.getTimeBetweenFrames() > 0
+                  ? 1 / direction.getTimeBetweenFrames()
+                  : 1,
+              name:
+                animationName ||
+                (resourceNames[0] &&
+                  removeExtensionFromFileName(resourceNames[0])) ||
+                objectName,
+              isLooping: direction.isLooping(),
+              existingMetadata: direction.getMetadata(),
+            },
+          }
+        );
+
+        setExternalEditorOpened(false);
+        if (!editResult) return;
+
+        const { resources, newMetadata, newName } = editResult;
+
+        const newDirection = new gd.Direction();
+        newDirection.setTimeBetweenFrames(direction.getTimeBetweenFrames());
+        newDirection.setLoop(direction.isLooping());
+        resources.forEach(resource => {
+          const sprite = new gd.Sprite();
+          sprite.setImageName(resource.name);
+          // Restore collision masks and points
+          if (
+            resource.originalIndex !== undefined &&
+            resource.originalIndex !== null
+          ) {
+            // The sprite existed before, so we can copy its points and collision masks.
+            const originalSprite = direction.getSprite(resource.originalIndex);
+            copySpritePoints(originalSprite, sprite);
+            copySpritePolygons(originalSprite, sprite);
+          } else {
+            // The sprite is new, apply points & collision masks if necessary.
+            applyPointsAndMasksToSpriteIfNecessary(
+              spriteConfiguration,
+              direction,
+              sprite
+            );
+          }
+          onSpriteAdded(sprite); // Call the callback before `addSprite`, as `addSprite` will store a copy of it.
+          newDirection.addSprite(sprite);
+          sprite.delete();
+        });
+
+        // Set metadata on the direction to allow editing again in the future.
+        if (newMetadata) {
+          newDirection.setMetadata(JSON.stringify(newMetadata));
+        }
+
+        // Burst the ResourcesLoader cache to force images to be reloaded (and not cached by the browser).
+        ResourcesLoader.burstUrlsCacheForResources(project, resourceNames);
+        replaceDirection(animationIndex, directionIndex, newDirection);
+
+        // If a name was specified in the external editor, use it for the animation.
+        if (newName) {
+          changeAnimationName(animationIndex, newName);
+        }
+        newDirection.delete();
+
+        if (onObjectUpdated) onObjectUpdated();
+        // If an external editor is used to edit the sprites, we assume the first sprite was edited.
+        if (animationIndex === 0) {
+          adaptCollisionMaskIfNeeded();
+        }
+      } catch (error) {
+        setExternalEditorOpened(false);
+        console.error(
+          'An exception was thrown when launching or reading resources from the external editor:',
+          error
+        );
+        showErrorBox({
+          message: `There was an error while using the external editor. Try with another resource and if this persists, please report this as a bug.`,
+          rawError: error,
+          errorId: 'external-editor-error',
+        });
+      }
+    },
+    [
+      spriteConfiguration,
+      project,
+      resourceManagementProps,
+      objectName,
+      replaceDirection,
+      onObjectUpdated,
+      onSpriteAdded,
+      changeAnimationName,
+      adaptCollisionMaskIfNeeded,
+    ]
+  );
+
+  const createAnimationWith = React.useCallback(
+    async (i18n: I18nType, externalEditor: ResourceExternalEditor) => {
+      addAnimation();
+      const direction = spriteConfiguration.getAnimation(0).getDirection(0);
+      await editWith(i18n, externalEditor, direction, 0, 0);
+    },
+    [addAnimation, editWith, spriteConfiguration]
   );
 
   const imageResourceExternalEditors = resourceManagementProps.resourceExternalEditors.filter(
@@ -532,11 +693,11 @@ export default function SpriteEditor({
                   <Trans>Animations are a sequence of images.</Trans>
                 }
                 actionLabel={<Trans>Import images</Trans>}
-                secondaryActionLabel={
-                  i18n._(isMobileScreen
+                secondaryActionLabel={i18n._(
+                  isMobileScreen
                     ? t`Draw`
-                    : imageResourceExternalEditors[0].createDisplayName)
-                }
+                    : imageResourceExternalEditors[0].createDisplayName
+                )}
                 secondaryActionIcon={<Edit />}
                 helpPagePath="/objects/sprite"
                 tutorialId="intermediate-changing-animations"
@@ -544,7 +705,14 @@ export default function SpriteEditor({
                   importImages();
                 }}
                 onSecondaryAction={
-                  imageResourceExternalEditors.length ? addAnimation : undefined
+                  imageResourceExternalEditors.length
+                    ? () => {
+                        createAnimationWith(
+                          i18n,
+                          imageResourceExternalEditors[0]
+                        );
+                    }
+                    : undefined
                 }
               />
             </Column>
@@ -668,6 +836,19 @@ export default function SpriteEditor({
                                           resourceManagementProps={
                                             resourceManagementProps
                                           }
+                                          editWith={(
+                                            i18n,
+                                            ResourceExternalEditor,
+                                            direction
+                                          ) =>
+                                            editWith(
+                                              i18n,
+                                              ResourceExternalEditor,
+                                              direction,
+                                              animationIndex,
+                                              directionIndex
+                                            )
+                                          }
                                           onReplaceByDirection={newDirection =>
                                             replaceDirection(
                                               animationIndex,
@@ -686,16 +867,7 @@ export default function SpriteEditor({
                                           onSpriteUpdated={onObjectUpdated}
                                           onFirstSpriteUpdated={
                                             animationIndex === 0
-                                              ? () => {
-                                                  // If the first sprite of the first animation is updated,
-                                                  // we update the automatic collision mask of the object,
-                                                  // if the option is enabled.
-                                                  if (
-                                                    spriteConfiguration.adaptCollisionMaskAutomatically()
-                                                  ) {
-                                                    onCreateMatchingSpriteCollisionMask();
-                                                  }
-                                                }
+                                              ? adaptCollisionMaskIfNeeded
                                               : undefined
                                           }
                                           onSpriteAdded={onSpriteAdded}
@@ -884,6 +1056,7 @@ export default function SpriteEditor({
               />
             </Dialog>
           )}
+          {externalEditorOpened && <ExternalEditorOpenedDialog />}
         </>
       )}
     </I18n>
